refactor(abw): decompose youtube2blog routes
